+++ projects/ngx-location-picker/src/lib/services/ngx-location-picker.helper.spec.ts
@@ -15,4 +15,40 @@ describe('NgxLocationPickerHelper', () => {
     it('should be created', () => {
         expect(service).toBeTruthy();
     });
+
+    describe('isAddressTests', () => {
+        const testCases = [
+            {
+                query: '',
+                expectedResult: false
+            },
+            {
+                query: 'Groenplaats',
+                expectedResult: false
+            },
+            {
+                query: 'Generaal armstrongweg',
+                expectedResult: false
+            },
+            {
+                query: 'Generaal armstrongweg 1',
+                expectedResult: true
+            },
+            {
+                query: 'Generaal armstrongweg 1A',
+                expectedResult: true
+            },
+            {
+                query: 'Kaainummer 757',
+                expectedResult: false
+            }
+        ];
+
+        testCases.forEach((test, index) => {
+            it(`should test if '${test.query}' isAddress with expectedResult: ${test.expectedResult}`, () => {
+                const result = service.isAddress(test.query);
+                expect(result).toEqual(test.expectedResult);
+            });
+        })
+    })
 });